WIP: profile icon moved, needs styling clean up
